Implemented edit task functionality

+++ js/main.js
@@ -1,4 +1,4 @@
-import { appendTasks, handleDeleteTask } from "./ui.js";
+import { appendTasks, handleDeleteTask, handleEditTask } from "./ui.js";
 import { addTask } from "./tasks.js";
 
 appendTasks();
@@ -13,7 +13,6 @@ document.getElementById("add-btn").addEventListener("click", () => {
 
 document.getElementById("tasks").addEventListener("click", (event) => {
   const targetBtn = event.target;
-  if (targetBtn.matches(".delete-btn")) {
-    handleDeleteTask(targetBtn);
-  }
+  if (targetBtn.matches(".delete-btn")) handleDeleteTask(targetBtn);
+  else if (targetBtn.matches(".edit-btn")) handleEditTask(targetBtn);
 });
